+++ medic-check-rust/src/main.rs
@@ -5,6 +5,7 @@ fn main() -> CheckResult {
     let cli = CliArgs::new();
 
     match cli.command {
+        Command::Audit(_) => medic_check_rust::cargo_audit()?,
         Command::CrateInstalled(args) => medic_check_rust::crate_installed(args.name)?,
         Command::FormatCheck(_) => medic_check_rust::check_formatting()?,
         Command::TargetInstalled(args) => medic_check_rust::target_installed(args.target)?,